@@ -1,218 +1,334 @@
-import React, { useState, useRef, useEffect } from 'react';
-import { NavLink } from "react-router-dom";
-import { FaPhoneAlt } from "react-icons/fa";
-import { IoMdMail } from "react-icons/io";
-import '../NavBar/index.css';
-import logoSRS from './../../assets/Images/srs/logo.png'
+import React, { useState } from 'react'
+import srsLogo from '../../assets/images/logo/srs.png';
+import { IoIosHelpBuoy } from "react-icons/io";
+import { ImBlog } from "react-icons/im";
+import { MdPhone, MdEventAvailable, MdOutlineAppSettingsAlt, MdDesignServices, MdOutlineDomainAdd } from "react-icons/md";
+import { CgWebsite } from "react-icons/cg";
+import { IoShieldCheckmarkOutline } from "react-icons/io5";
+import { FaDigitalOcean, FaBusinessTime } from "react-icons/fa6";
+import { SiJirasoftware } from "react-icons/si";
+import { BiSolidCctv } from "react-icons/bi";
+import './index.css';
 
 const Navbar = () => {
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-  const dropdownRef = useRef(null);
-  const buttonRef = useRef(null);
+    const [open, setOpen] = useState(false);
+    const [openSubmenu, setOpenSubmenu] = useState(null);
 
-  const toggleMenu = () => {
-    setIsMenuOpen(!isMenuOpen);
-  };
-
-  const toggleDropdown = (e) => {
-    e.stopPropagation();
-    setIsDropdownOpen(!isDropdownOpen);
-  };
-
-  const handleClickOutside = (e) => {
-    if (dropdownRef.current && !dropdownRef.current.contains(e.target) &&
-      buttonRef.current && !buttonRef.current.contains(e.target)
-    ) {
-      setIsDropdownOpen(false);
-    }
-  };
-
-  useEffect(() => {
-    document.addEventListener("mousedown", handleClickOutside);
-
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
+    const toggleSubMenu = (index) => {
+        setOpenSubmenu((prev) => (prev === index ? null : index));
     };
-  }, []);
 
-  const services = [
-    {
-      link: "/services/WebDevelopment",
-      name: "Web Development",
-      details: [
-        "Frontend Development",
-        "Backend Development",
-        "Full-stack Development",
-        "CMS Development",
-        "E-commerce Development",
-      ],
-    },
-    {
-      link: "/services/AppDevelopment",
-      name: "App Development",
-      details: [
-        "Native App Development",
-        "Cross-Platform Development",
-      ],
-    },
-    {
-      link: "/services/DigitalMarketing",
-      name: "Digital Marketing",
-      details: [
-        "Search Engine Optimization (SEO)",
-        "Social Media Marketing (SMM)",
-        "Content Marketing",
-        "Email Marketing",
-      ],
-    },
-    {
-      link: "/services/SoftwareDevelopment",
-      name: "Software Development",
-      details: [
-        "Custom Software Solutions",
-        "Enterprise Software Development",
-        "API Development and Integration",
-        "Desktop Application Development",
-      ],
-    },
-    {
-      link: "/services/UI&UXDesign",
-      name: "UI/UX Design",
-      details: [
-        "User Experience Research",
-        "Responsive Design",
-        "Interaction Design",
-        "Design Systems Development",
-      ],
-    },
-    {
-      link: "https://sysrootsolution.com/",
-      name: "Marketing CRM",
-      details: [
-        "Connect with third-party tools that you're already using.",
-      ],
-    },
-    {
-      link: "/services/CCTVInstallation",
-      name: "CCTV Installation",
-      details: [
-        "Residential CCTV Solutions",
-        "Commercial CCTV Solutions",
-        "CCTV Accessories",
-        "Additional Services",
-      ],
-    },
-    {
-      link: "/services/BusinessServices",
-      name: "Business Services",
-      details: [
-        "IT Support Services",
-        "Digital Transformation",
-        "Corporate Communication",
-        "Marketing and Branding",
-      ],
-    },
-    {
-      link: "/services/Domain&Hosting",
-      name: "Domain and Hosting",
-      details: [
-        "Domain Registration",
-        "Specialized Hosting",
-        "Domain and Hosting Management",
-        "Website Security Services",
-        "Technical Support",
-      ],
-    },
-  ];
+    const navLinks = [
+        {
+            title: "Home",
+            path: "/home",
+        },
+        {
+            title: "About",
+            path: "/about",
+        },
+        {
+            title: "Services",
+            subMenu: [
+                {
+                    logo: <MdOutlineAppSettingsAlt />,
+                    title: "App Development",
+                    subTitle: "Innovative and user-friendly mobile applications to elevate your business.",
+                    path: "",
+                },
+                {
+                    logo: <SiJirasoftware />,
+                    title: "Software Development",
+                    subTitle: "Custom software solutions for streamlined operations.",
+                    path: "",
+                },
+                {
+                    logo: <MdDesignServices />,
+                    title: "UX / UI Design",
+                    subTitle: "Crafting intuitive and visually appealing user experiences.",
+                    path: "",
+                },
+                {
+                    logo: <CgWebsite />,
+                    title: "Website Development",
+                    subTitle: "Responsive and dynamic websites tailored to your needs.",
+                    path: "",
+                },
+                {
+                    logo: <BiSolidCctv />,
+                    title: "CCTV Installation",
+                    subTitle: "Reliable surveillance solutions for your safety and security",
+                    path: "",
+                },
+                {
+                    logo: <FaDigitalOcean />,
+                    title: "Digital Marketing",
+                    subTitle: "Strategies that drive engagement and boost online presence.",
+                    path: "",
+                },
+                {
+                    logo: <MdOutlineDomainAdd />,
+                    title: "Domain & Hosting",
+                    subTitle: "Secure and scalable domain registration and hosting services.",
+                    path: "",
+                },
+                {
+                    logo: <FaBusinessTime />,
+                    title: "Business Services",
+                    subTitle: "Comprehensive solutions to support and grow your business.",
+                    path: "",
+                },
+            ],
+        },
+        {
+            title: "Products",
+            path: "/products",
+        },
+        {
+            title: "Clients",
+            path: "/clients",
+        },
+        {
+            title: "More",
+            subMenu: [
+                {
+                    logo: <ImBlog />,
+                    title: "Blogs",
+                    subTitle: "Discover More: Check Out Our Latest Blogs Here.",
+                    path: "",
+                },
+                {
+                    logo: <MdPhone />,
+                    title: "Contact Us",
+                    subTitle: "Need Assistance? Our Team is Just a Click Away!",
+                    path: "",
+                },
+                {
+                    logo: <MdEventAvailable />,
+                    title: "Events",
+                    subTitle: "Be Where It’s Happening! Explore Our Events Today.",
+                    path: "",
+                },
+                {
+                    logo: <IoIosHelpBuoy />,
+                    title: "Help Center",
+                    subTitle: "Need Assistance? Visit Our Help Center for Quick Solutions.",
+                    path: "",
+                },
+                {
+                    logo: <IoShieldCheckmarkOutline />,
+                    title: "Security",
+                    subTitle: "Understand how we take your privacy seriously.",
+                    path: "",
+                },
+            ],
+        },
+    ];
 
-  return (
-    <nav className="sticky top-0 z-[9999999]">
-      <div className='w-full bg-blue-100'>
-        <div className="flex flex-wrap justify-between items-center mx-auto max-w-screen-xl p-2 md:p-5 relative">
-          <a href="http://www.sysrootsolution.com" className="flex items-center space-x-3 rtl:space-x-reverse">
-            <img src={logoSRS} className="h-7 md:h-14 px-2" alt="SRS Logo" />
-          </a>
-          <button
-            onClick={toggleMenu}
-            type="button"
-            className="inline-flex items-center p-2 w-10 h-10 justify-center text-sm text-gray-500 rounded-lg md:hidden hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-gray-200 dark:text-gray-400 dark:hover:bg-gray-700 dark:focus:ring-gray-600"
-            aria-controls="mega-menu-full"
-            aria-expanded={isMenuOpen ? 'true' : 'false'}
-          >
-            <span className="sr-only">Open main menu</span>
-            <svg className="w-5 h-5" aria-hidden="true" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 17 14">
-              <path stroke="currentColor" strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M1 1h15M1 7h15M1 13h15" />
-            </svg>
-          </button>
-          <div className='hidden md:table bg-blue-50 py-1 px-3 shadow-lg cursor-pointer -rotate-0 absolute right-0 top-2 rounded mr-2 animate-bounce'>
-            <p className="flex items-center gap-2 text-gray-600 m-0"><FaPhoneAlt className='text-[#3F65B7]' /><span className='text-[20px] font-semibold'>+91-8789171115</span></p>
-            <p className='flex items-center gap-2 text-gray-600'><IoMdMail className='text-[#3F65B7]' />info@sysrootsolution.com</p>
-          </div>
-        </div>
-      </div>
-      <div className='flex flex-wrap justify-between items-center mx-auto w-full bg-[#3F65B7] nav-menu'>
-        <div
-          id="mega-menu-full"
-          className={`items-center justify-between mx-auto font-medium max-w-screen-xl w-full md:flex md:w-auto md:order-1 ${isMenuOpen ? 'block' : 'hidden'}`}
-        >
-          <ul className="flex flex-col p-2 md:p-0 rounded-lg md:space-x-8 rtl:space-x-reverse md:flex-row md:mt-0 md:border-0 dark:border-gray-700">
+    // Settings For mobile Navigation
+    const servicesSubMenu = navLinks.find(link => link.title === "Services")?.subMenu;
+    const moreSubMenu = navLinks.find(link => link.title === "More")?.subMenu;
+    const mainNavs = navLinks.filter(link => !link.subMenu);
+    const mobileNavs = [...mainNavs, ...(moreSubMenu || [])];
 
-            <li className='cursor-pointer'>
-              <NavLink to="/" className="text-lg block py-2 text-gray-100 rounded hover:text-[#3F65B7] hover:bg-blue-100 md:hover:bg-transparent md:hover:text-gray-300 md:p-0">Home</NavLink>
-            </li>
-            <li className='cursor-pointer'>
-              <NavLink to="/about" className="text-lg block py-2 text-gray-100 rounded hover:text-[#3F65B7] hover:bg-blue-100 md:hover:bg-transparent md:hover:text-gray-300 md:p-0">About</NavLink>
-            </li>
-            <li className='cursor-pointer'>
-              <span onClick={toggleDropdown} ref={buttonRef} className='text-lg block py-2 text-gray-100 rounded hover:text-[#3F65B7] hover:bg-gray-100 md:hover:bg-transparent md:hover:text-gray-300 md:p-0'>Services</span>
-            </li>
-            <li className='cursor-pointer'>
-              <NavLink to="/products" className="text-lg block py-2 text-gray-100 rounded hover:text-[#3F65B7] hover:bg-blue-100 md:hover:bg-transparent md:hover:text-gray-300 md:p-0">Products</NavLink>
-            </li>
-            <li className='cursor-pointer'>
-              <NavLink to="/clients" className="text-lg block py-2 text-gray-100 rounded hover:text-[#3F65B7] hover:bg-blue-100 md:hover:bg-transparent md:hover:text-gray-300 md:p-0">Clients</NavLink>
-            </li>
-            <li className='cursor-pointer'>
-              <NavLink to="/blog" className="text-lg block py-2 text-gray-100 rounded hover:text-[#3F65B7] hover:bg-blue-100 md:hover:bg-transparent md:hover:text-gray-300 md:p-0">Blog</NavLink>
-            </li>
-            <li className='cursor-pointer'>
-              <NavLink to="/contact" className="text-lg block py-2 text-gray-100 rounded hover:text-[#3F65B7] hover:bg-blue-100 md:hover:bg-transparent md:hover:text-gray-300 md:p-0">Contact</NavLink>
-            </li>
-            <li className='cursor-pointer'>
-              <NavLink to="/login" className="text-lg block py-2 underline text-blue-200 rounded hover:text-[#3F65B7] hover:bg-blue-100 md:hover:bg-transparent md:hover:text-emerald-100 md:p-0">Login</NavLink>
-            </li>
-          </ul>
-        </div>
-      </div>
-      {isDropdownOpen && (
-        <div id="mega-menu-full-dropdown" ref={dropdownRef} className="z-[99999] mt-1 border-gray-200 shadow bg-white dark:bg-gray-800  md:bg-blue-100 border-y dark:border-gray-600 static md:absolute w-full">
-          <div className="max-w-screen-xl px-4 py-5 mx-auto text-gray-900 dark:text-white md:px-6">
-            <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-0">
-              {services.map((service, index) => (
-                <ul key={index} className="space-y-2">
-                  <li>
-                    <NavLink
-                      to={service.link}
-                      className="block p-3 rounded-lg hover:bg-gray-50 dark:hover:bg-gray-700"
-                      onClick={() => setIsDropdownOpen(false)}
-                    >
-                      <div className="font-semibold text-[#3F65B7]">{service.name}</div>
-                      {service.details.map((detail, i) => (
-                        <p key={i} className="text-sm text-gray-500 dark:text-gray-400">
-                          {detail}
-                        </p>
-                      ))}
-                    </NavLink>
-                  </li>
-                </ul>
-              ))}
+    return (
+        <div className="relative bg-white">
+            <div className="max-w-7xl mx-auto px-4 sm:px-6">
+                <div className="flex justify-between items-center border-b-2 border-gray-100 py-3 lg:py-3 xl:py-4 xxl:py-5 lg:justify-start md:space-x-10">
+                    <div className="flex justify-start lg:w-0 lg:flex-1">
+                        <a href="#">
+                            <span className="sr-only">SRSolution</span>
+                            <img
+                                className="h-8 w-auto sm:h-10"
+                                src={srsLogo}
+                                alt=""
+                            />
+                        </a>
+                    </div>
+                    <div className="-mr-2 -my-2 lg:hidden">
+                        <button
+                            type="button"
+                            className="bg-white rounded-md p-2 inline-flex items-center justify-center text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500"
+                            onClick={() => setOpen(!open)}
+                        >
+                            <span className="sr-only">Open menu</span>
+                            <svg className="h-6 w-6 text-black" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
+                                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 6h16M4 12h16M4 18h16" />
+                            </svg>
+                        </button>
+                    </div>
+                    <nav className="hidden lg:flex space-x-6 lg:space-x-10">
+                        {
+                            navLinks.map((link, index) => {
+                                return (
+                                    <div key={index}>
+                                        {
+                                            link.path
+                                            &&
+                                            <a href="#" className="text-base font-medium text-gray-500 hover:text-srs-primary-hover">{link.title}</a>
+                                        }
+                                        {
+                                            link.subMenu
+                                            &&
+                                            <div className="relative">
+                                                <button type="button" className="group bg-white rounded-md text-gray-500 inline-flex items-center text-base font-medium hover:text-gray-900 "
+                                                    onClick={() => toggleSubMenu(index)}
+                                                >
+                                                    <span className='group-hover:text-srs-primary-hover'>{link.title}</span>
+                                                    <svg
+                                                        className={
+                                                            openSubmenu === index
+                                                                ? "transform rotate-180 ml-2 h-5 w-5 text-gray-400 group-hover:text-srs-primary-hover transition ease-out duration-200"
+                                                                : "ml-2 h-5 w-5 text-gray-400 group-hover:text-srs-primary-hover"
+                                                        }
+                                                        xmlns="http://www.w3.org/2000/svg"
+                                                        viewBox="0 0 20 20"
+                                                        fill="currentColor"
+                                                        aria-hidden="true"
+                                                    >
+                                                        <path
+                                                            fillRule="evenodd"
+                                                            d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z"
+                                                            clipRule="evenodd"
+                                                        />
+                                                    </svg>
+                                                </button>
+                                                <div
+                                                    onMouseLeave={() => setOpenSubmenu(null)}
+                                                    className={
+                                                        openSubmenu === index
+                                                            ? " opacity-100 translate-y-0 transition ease-out duration-200 absolute z-[999999] -ml-4 mt-3 transform px-2 w-screen max-w-md sm:px-0 lg:ml-0 lg:left-1/2 lg:-translate-x-1/2"
+                                                            : " opacity-0 translate-y-1 absolute z-[999999] -ml-4 mt-3 transform px-2 w-screen max-w-md sm:px-0 lg:ml-0 lg:left-1/2 lg:-translate-x-1/2"
+                                                    }
+                                                >
+                                                    <div className="rounded-lg shadow-lg ring-1 ring-black ring-opacity-5 overflow-hidden">
+                                                        <div className="relative grid gap-6 bg-white px-5 py-6 sm:gap-8 sm:p-8">
+
+                                                            {
+                                                                link.subMenu.map((link, index) => {
+                                                                    return (
+                                                                        <a
+                                                                            key={index}
+                                                                            href="#"
+                                                                            className="-m-3 p-2 flex items-start rounded-lg hover:bg-gray-50 group"
+                                                                        >
+                                                                            <span className="text-srs-primary linkLogo transform transition-transform duration-300 ease-in-out group-hover:scale-110">
+                                                                                {link.logo}
+                                                                            </span>
+                                                                            <div className="ml-4">
+                                                                                <p className="text-base font-medium text-gray-600">
+                                                                                    {link.title}
+                                                                                </p>
+                                                                                <p className="mt-1 text-xs text-gray-400">
+                                                                                    {link.subTitle}
+                                                                                </p>
+                                                                            </div>
+                                                                        </a>
+
+                                                                    )
+                                                                })
+                                                            }
+                                                        </div>
+                                                    </div>
+                                                </div>
+                                            </div>
+                                        }
+                                    </div>
+                                )
+                            })
+                        }
+                    </nav>
+                    <div className="hidden lg:flex items-center justify-end md:flex-1 lg:w-0">
+                        <a href="#" className="ml-8 whitespace-nowrap inline-flex items-center justify-center px-4 py-2 border border-transparent rounded-md shadow-sm text-base font-medium text-white bg-srs-primary hover:bg-srs-primary-hover">
+                            Sign In
+                        </a>
+                    </div>
+                </div>
             </div>
-          </div>
-        </div>
-      )}
-    </nav>
-  );
-};
 
-export default Navbar;
+            {/* Mobile Navigation */}
+            <div
+                className={
+                    open
+                        ? "opacity-100 scale-100 ease-out duration-200 absolute top-0 inset-x-0 p-2 transition transform origin-top-right lg:hidden z-[999999]"
+                        : "opacity-0 scale-95 absolute top-0 inset-x-0 p-2 transition transform origin-top-right lg:hidden z-[999999]"
+                }
+            >
+                <div className="rounded-lg shadow-lg ring-1 ring-black ring-opacity-5 bg-white divide-y-2 divide-gray-50">
+                    <div className="pt-5 pb-6 px-5 bg-blue-50">
+                        <div className="flex items-center justify-between">
+                            <div>
+                                <img className="h-8 w-auto" src={srsLogo} alt="SRSolution" />
+                            </div>
+                            <div className="-mr-2">
+                                <button type="button" className="rounded-md p-1 inline-flex items-center justify-center text-gray-400 hover:text-srs-primary-hover"
+                                    onClick={() => setOpen(!open)}
+                                >
+                                    <span className="sr-only">Close menu</span>
+                                    <svg
+                                        className="h-6 w-6"
+                                        xmlns="http://www.w3.org/2000/svg"
+                                        fill="none"
+                                        viewBox="0 0 24 24"
+                                        stroke="currentColor"
+                                        aria-hidden="true"
+                                    >
+                                        <path
+                                            strokeLinecap="round"
+                                            strokeLinejoin="round"
+                                            strokeWidth={2}
+                                            d="M6 18L18 6M6 6l12 12"
+                                        />
+                                    </svg>
+                                </button>
+                            </div>
+                        </div>
+                        <div className="border-white rounded-md border-2 p-8 mt-4 bg-white">
+                            <nav className="grid gap-y-5">
+                                {servicesSubMenu?.map((item, index) => (
+                                    <div key={index} >
+                                        <a
+                                            href="#"
+                                            className="-m-2 px-3 py-1 flex items-center rounded-md hover:bg-gray-50"
+                                        >
+                                            <span className='linkLogo text-srs-primary'>{item.logo}</span>
+                                            <span className="ml-3 text-base font-medium text-gray-500">
+                                                {item.title}
+                                            </span>
+                                        </a>
+                                    </div>
+                                ))}
+                            </nav>
+                        </div>
+                    </div>
+                    <div className="py-6 px-5 space-y-6">
+                        <div className="grid grid-cols-2 gap-y-4 gap-x-8  bg-blue-50 rounded-md p-5">
+                            {mobileNavs?.map((item, index) => (
+                                <div key={index}>
+                                    <a
+                                        href="#"
+                                        className="-m-4 px-3 py-2 flex items-center rounded-md text-srs-primary"
+                                    >
+                                        <span className="ml-3 text-base font-medium ">
+                                            {item.title}
+                                        </span>
+                                    </a>
+                                </div>
+                            ))}
+                        </div>
+                        <div>
+                            <a
+                                href="#"
+                                className="w-full flex items-center justify-center px-4 py-2 border border-transparent rounded-md shadow-sm text-base font-medium text-white bg-srs-primary hover:bg-srs-primary-hover"
+                            >
+                                Sign In
+                            </a>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    )
+}
+
+export default Navbar
